requirements.txt
flask
flask-login
flask-restful
flask-mail
flask-wtf
flask-sqlalchemy
sqlalchemy-serializer
